requirements.txt
Pillow==11.3.0
pvporcupine==3.0.5
PyAudio==0.2.14
pyautogui==0.9.54
python-dotenv==1.1.1
Requests==2.32.5
streamlit==1.49.0
websocket_client==1.8.0
pytest==8.3.3
pytest-cov==5.0.0